usbredirect: Introduce usbredir TCP client
There was an old request [0] to make an usbredir client that would be
similar to usbredirserver but working as a TCP client instead. This
way, if a host like QEMU is configured with usb-redir, an user can use
this tool to redirect a USB device in the client machine.

I'm using glib/gio to make this tool portable to Windows too. I've
tested and works fine in Windows 10 with UsbDk installed [1].

The following patch is going to extend usbredirect tool to act as
server as well, so we could have a single tool acting as TCP client
and server in Unix and Windows systems.

[0] https://bugzilla.redhat.com/show_bug.cgi?id=844657
[1] https://github.com/daynix/UsbDk

Tested this patch with the following libvirt configuration:

    <redirdev bus='usb' type='tcp'>
        <source mode='bind' host='localhost' service='5550'/>
        <protocol type='raw'/>
        <address type='usb' bus='0' port='3'/>
    </redirdev>

Signed-off-by: Victor Toso <victortoso@redhat.com>