usbredirect: Add keepalive flag
In order to make feature parity with usbredirserver,
enabling/disabling SO_KEEPALIVE is necessary.

GIO does not have an API to change default value (which depends on
protocol, etc). If this seems necessary in the future, we either add
that to GIO or retrieve the fd and do set it ourselves.

Signed-off-by: Victor Toso <victortoso@redhat.com>