build-sys: add meson option for fuzzing
Compiler's LLVMFuzzerTestOneInput is provided (llvm >= 11.0.0), it'll
use that; otherwise, we build the standalone static library for our
usbredirparserfuzz binary.

There is no `fuzzer/FuzzedDataProvider.h` for windows, so we have it
disabled by default for that host system.

As the overall settings around libFuzzer relies on environment
variables, we add build-aux/oss-fuzz.sh script which is based on
systemd's script [0], hence, I'm keeping the name here.

    [0] https://github.com/systemd/systemd/blob/main/tools/oss-fuzz.sh

Signed-off-by: Victor Toso <victortoso@redhat.com>