build-sys: Add msi installer generation
Using wixl and wixl-heat from msitools package, we can generate a msi
installer for usbredirect binary.

The user needs the mingw-glib and mingw-libusbx dependencies for
either x64 or i686. The process to generate the installer is:

  $ mkdir build-win64 && cd build-win64
  $ mingw64-meson
  $ DESTDIR=./install-root ninja install
  $ DESTDIR=./install-root ninja data/usbredirect-x64-0.8.0.msi

Fixes: https://gitlab.freedesktop.org/spice/usbredir/-/issues/12

Note that this patch takes lots of its logic from virt-viewer msi
generator code including build-aux/msitool.py [0] which was only
adapted to fit usbredir here.

[0] https://gitlab.com/virt-viewer/virt-viewer/-/blob/master/build-aux/msitool.py

Signed-off-by: Victor Toso <victortoso@redhat.com>