Use single stateless function for filter tests
Recent commits have added unittests for the filter parsing code. The test code
contained two similar, but different implementations of table-driven tests. With
the way the code was written it was rather hard to identify which exact case
failed.

With this change every test case runs under its own "testpath", making it
recognizable from the output (e.g. "ok 28 /filter/rules/bad/17"). In addition
the evaluated filter is emitted in escaped form as a debug message.

The previous implementation of the test for acceptable filter rules was
stateful, i.e. changing the token or rule separator would have an effect on
later test cases. In order to have a predictable test situation it's better
to always fully specify a full test case.

Signed-off-by: Michael Hanselmann <public@hansmi.ch>