usbredirparserfuzz: Use constant for magic number
Now that the magic number for the serialization format is a header it can also
be used for fuzzing. As it turned out I had gotten the endianess wrong in commit
58f198e and the unserialization code wasn't actually fuzzed (unless the fuzzer
had somehow found the magic value).

Endianess support is still in the TODO file and so a plain memcpy is sufficient.

Signed-off-by: Michael Hanselmann <public@hansmi.ch>